+++ feature/profile/create/src/commonMain/kotlin/com/whatever/caramel/feature/profile/create/mvi/ProfileCreateState.kt
@@ -1,7 +1,7 @@
 package com.whatever.caramel.feature.profile.create.mvi
 
 import com.whatever.caramel.core.domain.vo.user.Gender
-import com.whatever.caramel.core.ui.picker.DateUiState
+import com.whatever.caramel.core.ui.picker.model.DateUiState
 import com.whatever.caramel.core.viewmodel.UiState
 
 data class ProfileCreateState(
